add robolectric integration tests
