+++ build.gradle
@@ -7,7 +7,7 @@ buildscript {
     ext.ktlintVersion = '0.36.0'
     ext.ktlintGradlePluginVersion = '9.2.1'
     ext.junitVersion = '4.12'
-    ext.robolectricVersion = '4.1'
+    ext.robolectricVersion = '4.3'
     ext.rxJavaVersion = '3.0.3'
     ext.truthVersion = '0.40'
