+++ overlapping_panels/build.gradle
@@ -31,11 +31,4 @@ dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
     implementation "androidx.appcompat:appcompat:$androidxAppCompatVersion"
     implementation "androidx.core:core-ktx:$androidxCoreVersion"
-
-    androidTestImplementation "junit:junit:$junitVersion"
-    androidTestImplementation "com.google.truth:truth:$truthVersion"
-    androidTestImplementation("org.robolectric:robolectric:$robolectricVersion") {
-        exclude group: "org.apache.maven.wagon"
-        exclude group: "org.apache.maven"
-    }
 }
