--- overlapping_panels/src/androidTest/res/layout/overlapping_panels_test_layout.xml
@@ -1,48 +1,0 @@
-<com.discord.panels.OverlappingPanelsLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/overlapping_panels"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <FrameLayout
-        android:id="@+id/start_panel"
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent"
-        android:layout_gravity="start">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:text="Start Panel" />
-
-    </FrameLayout>
-
-    <FrameLayout
-        android:id="@+id/center_panel"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:text="Center Panel" />
-
-    </FrameLayout>
-
-    <FrameLayout
-        android:id="@+id/end_panel"
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent"
-        android:layout_gravity="end">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:text="End Panel" />
-
-    </FrameLayout>
-
-</com.discord.panels.OverlappingPanelsLayout>
