+++ sample_app/src/main/res/layout/main_activity.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".MainActivity">
+    tools:context="com.discord.sampleapp.MainActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:layout_width="match_parent"
